Final commit

+++ student-catalog-service/src/main/java/com/ironhack/student_catalog_service/models/StudentGrade.java
@@ -16,6 +16,6 @@ public class StudentGrade {
 
     private String name;
     private int age;
-    private int grade;
+    private float grade;
 
 }
